+++ src/fit/include/RAT/BONSAI/goodness.h
@@ -16,13 +16,13 @@ class goodness : public fit_param, public fitquality, public hitsel {
   float *ttof, *tau, *hitdir;                      // arrays for time residuals and directions
   float t0, cosc, theta, phi, gdns, gdn0;          // fit results
   float twin, tcent, tnorm;                        // time window, ave. time, time normal.
-  short int *lists, *buffer, *sort_list, nbranch;  // for sorting
-  short int ngdn;                                  // number of fit vertices
+  int *lists, *buffer, *sort_list, nbranch;  // for sorting
+  int ngdn;                                  // number of fit vertices
   dodecahedron dod;                                // defines how surrounding points
   axes orientation;                                // are calculated
 
   inline void create_lists(int n);           // create hitsel sorting lists and buffer
-  void qsort(short int *list, short int n);  // recursive QuickSort program
+  void qsort(int *list, int n);  // recursive QuickSort program
   void bubble(void);                         // Bubble sort program
 
  public:
@@ -46,14 +46,14 @@ class goodness : public fit_param, public fitquality, public hitsel {
   inline virtual int nresult(void);                 // return size of result array
   inline virtual void get_result(float *r);         // get result array
   inline virtual void set_result(float *r);         // set result array
-  inline virtual void set_branch(short int point);  // switch memory to search branch
+  inline virtual void set_branch(int point);  // switch memory to search branch
   // maximizing procedure
   inline void maximize(bonsaifit *fit, searchgrid *grid, bool useAngle = true);
   inline void maximize(bonsaifit *cfit, bonsaifit *bfit, likelihood *bslike, searchgrid *grid, bool useAngle = true);
   inline float get_zero(void);           // get t0
   inline float get_gdn(void);            // get goodness
   inline float get_gdn0(void);           // get goodness
-  inline float residual(short int hit);  // get time residual
+  inline float residual(int hit);  // get time residual
   inline float get_theta(void);          // get zenith angle
   inline float get_phi(void);            // get azimuth angle
   inline float get_cos_theta(void);      // get opening angle
